[REL] migration product_variant_o2o to 11.0 * [FIX] Update Readme * [IMP] create variant function with one variant per product * [FIX] remove data tags and api.one decorator * [FIX] update version module
