+++ product_variant_o2o/models/__init__.py
@@ -2,4 +2,4 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import models
+from . import product_template
