+++ product_variant_o2o/views/product_template_views.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="product_template_only_form_view" model="ir.ui.view">
+        <field name="name">product.template.form.inherit</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_only_form_view"></field>
+        <field name="arch" type="xml">
+            <p class="oe_grey" position="before">
+                <label for="one_variant_per_product"/>
+                <field name="one_variant_per_product"/>
+            </p>
+            <p class="oe_grey" position="attributes">
+                <attribute name="attrs">{'invisible': [('one_variant_per_product','=',True)]}</attribute>
+            </p>
+            <p class="oe_grey" position="after">
+                <p class="oe_grey" attrs="{'invisible': [('one_variant_per_product','=',False)]}">
+                    <strong>Warning</strong>: Change attributes won't change variants, it will only update variants attribute.
+                </p>
+            </p>
+        </field>
+    </record>
+</odoo>
